+++ src/plugins/search/plugin.py
@@ -18,8 +18,6 @@
 # FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
 # IN THE SOFTWARE.
 
-import re
-
 from html import escape
 from html.parser import HTMLParser
 from mkdocs.contrib.search import SearchPlugin as BasePlugin
@@ -42,15 +40,8 @@ class SearchPlugin(BasePlugin):
 # Search index with support for additional fields
 class SearchIndex(BaseIndex):
 
+    # A simple wrapper to add an entry, dropping bad characters
     def _add_entry(self, title, text, loc):
-        """
-        A simple wrapper to add an entry, dropping bad characters.
-        """
-
-        # text = text.replace('\u00a0', ' ')
-        # text = re.sub(r'[ \t\n\r\f\v]+', ' ', text.strip())
-        text = re.sub(r'\n\n+', '\n\n', text).strip()
-
         self._entries.append({
             'title': title,
             'text': text,
@@ -59,7 +50,7 @@ class SearchIndex(BaseIndex):
 
     # Override to add additional fields for each page
     def add_entry_from_context(self, page):
-        index = len(self._entries)
+        #index = len(self._entries)
         # super().add_entry_from_context(page)
 
         ###
@@ -71,37 +62,25 @@ class SearchIndex(BaseIndex):
         parser.feed(page.content)
         parser.close()
 
-        #print(parser.stripped_html)
-
         # Get the absolute URL for the page, this is then
         # prepended to the urls of the sections
         url = page.url
-
-        # Create an entry for the full page.
-        text = parser.stripped_html.rstrip('\n') if self.config['indexing'] == 'full' else ''
-        self._add_entry(
-            title=page.title,
-            text=text,
-            loc=url
-        )
-
         if self.config['indexing'] in ['full', 'sections']:
             for section in parser.data:
-                self.create_entry_for_section(section, page.toc, url)
-        ###
+                self.create_entry_for_section(section, page.toc, url, page)
 
-        entry = self._entries[index]
+        # entry = self._entries[index]
 
-        # Add document tags
-        if "tags" in page.meta:
-            entry["tags"] = page.meta["tags"]
+        # # Add document tags
+        # if "tags" in page.meta:
+        #     entry["tags"] = page.meta["tags"]
 
-        # Add document boost for search
-        search = page.meta.get("search", {})
-        if "boost" in search:
-            entry["boost"] = search["boost"]
+        # # Add document boost for search
+        # search = page.meta.get("search", {})
+        # if "boost" in search:
+        #     entry["boost"] = search["boost"]
 
-    def create_entry_for_section(self, section, toc, abs_url):
+    def create_entry_for_section(self, section, toc, abs_url, page):
         """
         Given a section on the page, the table of contents and
         the absolute url for the page create an entry in the
@@ -111,25 +90,20 @@ class SearchIndex(BaseIndex):
         toc_item = self._find_toc_by_id(toc, section.id)
 
         text = ''.join(section.text) if self.config['indexing'] == 'full' else ''
-        if toc_item is not None:
+        # TODO: when literal h1, h2 etc are used, this won't work
+        if toc_item is not None and section.tag != "h1":
             self._add_entry(
-                title=toc_item.title,
+                title="".join(section.title),
                 text=text,
                 loc=abs_url + toc_item.url
             )
-
-    # def _add_entry(self, title, text, loc):
-    #     """
-    #     A simple wrapper to add an entry, dropping bad characters.
-    #     """
-    #     text = text.replace('\u00a0', ' ')
-    #     # text = re.sub(r'[ \t\n\r\f\v]+', ' ', text.strip())
-    #     # print(text)
-    #     self._entries.append({
-    #         'title': title,
-    #         'text': text,
-    #         'location': loc
-    #     })
+        else:
+            # this is a whole page entry!
+            self._add_entry(
+                title=page.title,
+                text=text,
+                loc=abs_url
+            )
 
 class ContentSection:
     """
@@ -137,10 +111,11 @@ class ContentSection:
     need when it is parsing the HMTL.
     """
 
-    def __init__(self, text=None, id_=None, title=None):
+    def __init__(self, tag, text=None, id_=None, title=None):
         self.text = text or []
         self.id = id_
-        self.title = title
+        self.title = title or []
+        self.tag = tag
 
     def __eq__(self, other):
         return (
@@ -159,127 +134,89 @@ class ContentParser(HTMLParser):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
+        # Tags to filter (i.e. to not include contents in index)
+        self.filter = set([
+            "img",
+            "object",
+            "script",
+            "style"
+        ])
+
+        # Tags to preserve
+        self.preserve = set([
+            "code",
+            "li",
+            "ol",
+            "p",
+            "pre",
+            "ul",
+            # "table",
+            # "td",
+            # "th",
+            # "tr"
+        ])
+
+        self.stack = []
+
         self.data = []
         self.section = None
 
-        self.is_header_tag = False
-        self.is_pre_tag = None
-        self.is_script_tag = False
-
-        self._stripped_html = []
-
+    # Called at the start of every HTML tag
     def handle_starttag(self, tag, attrs):
-        """Called at the start of every HTML tag."""
+        self.stack.append(tag)
 
-        # We only care about the opening tag for headings.
+        # Handle headings
         if tag in ([f"h{x}" for x in range(1, 7)]):
-            # We are dealing with a new header, create a new section
-            # for it and assign the ID if it has one.
-            self.is_header_tag = True
-            self.section = ContentSection()
+            self.section = ContentSection(tag)
             self.data.append(self.section)
 
+            # Set identifier on section for TOC resolution
             for attr in attrs:
                 if attr[0] == "id":
                     self.section.id = attr[1]
 
-            return
+        # Handle preface to headings - ensure top-level section
+        if not self.section:
+            self.section = ContentSection("h1")
+            self.data.append(self.section)
 
-        a = dict(attrs)
-        if tag == "a" and a.get("class", None) == "headerlink":
-            self.is_script_tag = True
-            return
+        # Render opening tag if preserved
+        if tag in self.preserve:
+            text = self.section.text
+            if self.section.tag in self.stack:
+                text = self.section.title
 
-        # OVERRIDE
-        # if tag == "div":
-        #     self.is_pre_tag = "div"
+            # Append to section title or text
+            text.append("<{}>".format(tag))
 
-        # if tag == "ul":
-        #     self.is_pre_tag = "ul"
-
-        # if tag == "ol":
-        #     self.is_pre_tag = "ol"
-
-        # if tag == "pre":
-        #     self.is_pre_tag = "pre"
-
-        # if tag == "code" and not self.is_pre_tag:
-        #     self.is_pre_tag = "code"
-
-        self.is_pre_tag = True
-
-        if tag == "script" or tag == "img":
-            self.is_script_tag = True
-            return
-
-        # if self.is_pre_tag and self.section:
-        #     more = ["{}='{}'".format(k, v) for k, v in attrs]
-        #     if tag == "pre":
-        #         more.append("class=highlight")
-        #     self.section.text.append("<{} {}>".format(tag, " ".join(more)))
-
-        if self.section:
-            more = ["{}='{}'".format(k, v) for k, v in attrs]
-            if tag == "pre":
-                more.append("class=highlight")
-            self.section.text.append("<{} {}>".format(tag, " ".join(more)))
-
-
+    # Called at the end of every HTML tag
     def handle_endtag(self, tag):
-        """Called at the end of every HTML tag."""
+        if self.stack[-1] == tag:
+            self.stack.pop()
 
-        # We only care about the opening tag for headings.
-        if tag in ([f"h{x}" for x in range(1, 7)]):
-            self.is_header_tag = False
-            return
+        # Render closing tag if preserved
+        if tag in self.preserve:
+            text = self.section.text
+            if self.section.tag in self.stack:
+                text = self.section.title
 
-        if tag == "script" or tag == "img" or (tag == "a" and self.is_script_tag):
-            self.is_script_tag = False
-            return
+            # Append to section title or text
+            text.append("</{}>".format(tag))
 
-        # OVERRIDE
-        # if self.is_pre_tag and self.section:
-        #     self.section.text.append("</{}>".format(tag))
-
-        if self.section:
-            self.section.text.append("</{}>".format(tag))
-
-        # TODO: better use a stack
-        if tag == self.is_pre_tag:
-            self.is_pre_tag = None
-
+    # Called for the text contents of each tag.
     def handle_data(self, data):
-        """
-        Called for the text contents of each tag.
-        """
-
-        if self.is_script_tag:
+        if self.filter.intersection(self.stack):
             return
 
-        # if self.is_pre_tag:
-        #     self._stripped_html.append(escape(data))
-        # else:
-        #     self._stripped_html.append(data)#
+        # Collapse whitespace in non-pre contexts
+        if not "pre" in self.stack:
+            data = data.strip("\n").replace("\n", " ")
 
-        # TODO: MINIFY HTML!!!
+        # Ignore section headline
+        if self.section.tag in self.stack:
+            if not "a" in self.stack:
+                self.section.title.append(escape(data))
 
-        if self.section is None:
-            # This means we have some content at the start of the
-            # HTML before we reach a heading tag. We don't actually
-            # care about that content as it will be added to the
-            # overall page entry in the search. So just skip it.
-            return
-
-        # If this is a header, then the data is the title.
-        # Otherwise it is content of something under that header
-        # section.
-        if self.is_header_tag:
-            self.section.title = data
-        elif self.is_pre_tag:
-            self.section.text.append(escape(data))
+        # Handle everything else
         else:
-            self.section.text.append(data)
-
-    @property
-    def stripped_html(self):
-        return ''.join(self._stripped_html)
+            self.section.text.append(escape(data))