+++ docs/changelog.md
@@ -8,7 +8,12 @@ foobar lol fuck you
 
 Here's some text after the changelog!
 
-### 7.2.0 <small>_ July 25, 2021</small>
+### 7.2.2 <small>_ July 31, 2021</small>
+
+- Updated Korean translations
+- Fixed #2879: Search highlighting does not properly escape HTML
+
+### 7.2.1 <small>_ July 25, 2021</small>
 
 - Fixed #2862: Back-to-top button overlays active search bar